llm_ops_project/requirements.txt
langchain>=0.0.200
openai>=0.27.0
pinecone-client>=2.2.1
pyyaml>=6.0
python-dotenv>=0.19.0
asyncio>=3.4.3
pytest>=7.0.0
black>=22.3.0
isort>=5.10.1
flake8>=4.0.1
jupyter>=1.0.0 
boto3>=1.26.0 
kfp>=1.8.0 
prometheus-client>=0.16.0 
watchtower>=3.0.0 
scikit-learn>=1.0.0
matplotlib>=3.4.0
pandas>=1.3.0
transformers #TODO: do we need a version
langchain-community #TODO: do we need a version
sentence-transformers #TODO: do we need a version